requirements.txt
streamlit
google-genai
python-dotenv
streamlit-ace
build
pylint
pyyaml